+++ lib/kitsu.dart
@@ -1,0 +1,12 @@
+/**
+ * Kitsumon - A fox-powered Kitsu.io API Wrapper for Dart.
+ * Copyright (c) 2020 Nebulino
+ */
+
+/// Kitsu Library.
+/// It contains different types you can find using [KitsuAPI].
+///
+/// [KitsuAPI]: https://kitsu.docs.apiary.io/
+library kitsu;
+
+export 'package:kitsumon/src/kitsu.dart';