+++ lib/kitsu_methods.dart
@@ -1,0 +1,9 @@
+/**
+ * Kitsumon - A fox-powered Kitsu.io API Wrapper for Dart.
+ * Copyright (c) 2020 Nebulino
+ */
+
+/// It contains and group up all the methods.
+library kitsu_methods;
+
+export 'package:kitsumon/src/methods/authentication_methods.dart';