+++ lib/kitsumon_exceptions.dart
@@ -1,0 +1,11 @@
+/**
+ * Kitsumon - A fox-powered Kitsu.io API Wrapper for Dart.
+ * Copyright (c) 2020 Nebulino
+ */
+
+/// Kitsumon Exception Library.
+/// It contains some of the exceptions of my library.
+library scrobblenaut_exceptions;
+
+export 'package:kitsumon/src/exceptions/api_exception.dart';
+export 'package:kitsumon/src/exceptions/kitsumon_exception.dart';