+++ lib/src/kitsu.dart
@@ -1,0 +1,15 @@
+/**
+ * Kitsumon - A fox-powered Kitsu.io API Wrapper for Dart.
+ * Copyright (c) 2020 Nebulino
+ */
+
+/// Kitsu Type Class.
+library kitsu_objects;
+
+import 'package:json_annotation/json_annotation.dart';
+import 'package:meta/meta.dart';
+import 'package:kitsumon/src/helpers/kitsu_value_normalizer.dart';
+
+part 'kitsu.g.dart';
+
+part 'package:kitsumon/src/kitsu/authentication.dart';