+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.0.2
+#### Base Implementation of Kitsumon.
+
+- Created Kitsu object.
+- Created API and Kitsumon exceptions.
+- Created KitsuValueNormalizer helper.
+- Created Authentication Method Class. (The only available one, more in the future?)
+- Created KitsuClient.
+- Created first KitsuObject, Authentication.
+- Created base implementation of Kitsumon.
+- Added build and clean Scripts.
+- Updated the README.md after talking with Kitsu developers.
+
 ## 0.0.1
 
 - Initial version. Created the Package's repo structure.